pyproject.toml
[project]
name = "professional-claim-structurer"
version = "0.1.0"
description = "Generalized structured claim generation for professional decision support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12,<5.0",
    "httpx>=0.27,<1.0",
    "openai>=1.99.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "python-dotenv>=1.0,<2.0",
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.30,<1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
